+++ src/mympd_client/partitions.h
@@ -17,6 +17,5 @@ bool partitions_connect(struct t_mympd_state *mympd_state, struct t_partition_st
 struct t_partition_state *partitions_get_by_name(struct t_mympd_state *mympd_state, const char *name);
 void partitions_list_clear(struct t_mympd_state *mympd_state);
 bool partitions_populate(struct t_mympd_state *mympd_state);
-bool partitions_check(struct t_mympd_state *mympd_state, const char *name);
-void partitions_add(struct t_mympd_state *mympd_state, const char *name);
+
 #endif